IT (sort of) WORKS!
I finally got round to implementing everything needed to get sectors working. Added colours and of course the Intersecting New Vector algorithm. Still some really awful bugs that need fixing before it is really usable but just so happy that I have got this far!

It's starting to look like a Map Editor now which is a good start