Rework and starting debug
Just starting to get a handle on the code after a period of absence. My next step is to mostly rewrite the map class with more of a focus for edge cases which were hard to implement before hand. The errors that are coming back now are based on conflicts between ComputerVectors and UserVectors